Ensure that navigation headings have unique IDs
Previously, the heading elements (with class .panel-heading) in the
navigation menu all had the same ID, #headingOne. This is an
accessibility concern, as duplicate IDs may cause screen readers to skip
elements.

This commit fixes this issue by assigning a unique ID to each such
element, based on the for loop counter.

This improves the Accessibility score in Lighthouse from 82 to 89, when
testing on the home page in desktop mode.